correcao 25

+++ src/main/java/com/geracaocaldeira/java/Aula13/Exercicio29/MinibossExemplo/Model/Livro.java
@@ -10,14 +10,14 @@ public class Livro {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String titulo;
     private String autor;
     private String isbn;
-    private int ano;
+    private Integer ano;
 
-    public Long getId() {return id;}
-    public void setId(Long id) {this.id = id;}
+    public Integer getId() {return id;}
+    public void setId(Integer id) {this.id = id;}
     public String getTitulo() {return titulo;}
     public void setTitulo(String titulo) {this.titulo = titulo;}
     public String getAutor() {return autor;}
@@ -25,6 +25,6 @@ public class Livro {
     public String getIsbn() {return isbn;}
     public void setIsbn(String isbn) {this.isbn = isbn;}
     public int getAno() {return ano;}
-    public void setAno(int ano) {this.ano = ano;}
+    public void setAno(Integer ano) {this.ano = ano;}
 }
